// src/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Address = Node & {
  __typename?: 'Address';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  fullName?: Maybe<Scalars['String']>;
  company?: Maybe<Scalars['String']>;
  streetLine1: Scalars['String'];
  streetLine2?: Maybe<Scalars['String']>;
  city?: Maybe<Scalars['String']>;
  province?: Maybe<Scalars['String']>;
  postalCode?: Maybe<Scalars['String']>;
  country: Country;
  phoneNumber?: Maybe<Scalars['String']>;
  defaultShippingAddress?: Maybe<Scalars['Boolean']>;
  defaultBillingAddress?: Maybe<Scalars['Boolean']>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Channel = Node & {
  __typename?: 'Channel';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  code: Scalars['String'];
  token: Scalars['String'];
  defaultTaxZone?: Maybe<Zone>;
  defaultShippingZone?: Maybe<Zone>;
  defaultLanguageCode: LanguageCode;
  currencyCode: CurrencyCode;
  pricesIncludeTax: Scalars['Boolean'];
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type CreateNfeCompanyInput = {
  ownerName: Scalars['String'];
  corporateName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  tradeName?: Maybe<Scalars['String']>;
  accounting?: Maybe<Scalars['String']>;
  subdomain?: Maybe<Scalars['String']>;
  logoUrl?: Maybe<Scalars['String']>;
  cpf: Scalars['String'];
  cnpj: Scalars['String'];
  ie: Scalars['String'];
  unit: NfeCompanyUnit;
  address: NfeCompanyAddressInput;
  taxSettings?: Maybe<NfeCompanyTaxSettingsInput>;
  certificate?: Maybe<NfeCompanyCertificateInput>;
  emitterCredentials?: Maybe<NfeCompanyEmitterCredentialsInput>;
};
// ...
export type Nfe = Node & {
  __typename?: 'Nfe';
  id: Scalars['ID'];
  createdAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
  state?: Maybe<Scalars['String']>;
  env: NfeEnv;
  series?: Maybe<Scalars['String']>;
  number?: Maybe<Scalars['String']>;
  receipt?: Maybe<Scalars['String']>;
  key?: Maybe<Scalars['String']>;
  xml?: Maybe<Scalars['String']>;
  danfe?: Maybe<Scalars['String']>;
  info: NfeInfo;
  recipient: NfeRecipient;
  payment: NfePayment;
  transport?: Maybe<NfeTransport>;
  exportation?: Maybe<NfeExportation>;
  references?: Maybe<NfeReferences>;
  company: NfeCompany;
  order: Order;
  lastOperation?: Maybe<NfeOperation>;
  operations?: Maybe<Array<Maybe<NfeOperation>>>;
  products?: Maybe<Array<NfeProduct>>;
};

export type NfeCompany = Node & {
  __typename?: 'NfeCompany';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  ownerName: Scalars['String'];
  corporateName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  tradeName?: Maybe<Scalars['String']>;
  accounting?: Maybe<Scalars['String']>;
  subdomain?: Maybe<Scalars['String']>;
  logoUrl?: Maybe<Scalars['String']>;
  unit: NfeCompanyUnit;
  cpf: Scalars['String'];
  cnpj: Scalars['String'];
  ie: Scalars['String'];
  address: NfeCompanyAddress;
  taxSettings?: Maybe<NfeCompanyTaxSettings>;
  certificate?: Maybe<NfeCompanyCertificate>;
  emitterCredentials?: Maybe<NfeCompanyEmitterCredentials>;
  channels?: Maybe<Array<Maybe<Channel>>>;
  nfes?: Maybe<Array<Maybe<Nfe>>>;
};

export type NfeCompanyAddress = {
  __typename?: 'NfeCompanyAddress';
  postcode: Scalars['String'];
  street: Scalars['String'];
  number: Scalars['String'];
  complement?: Maybe<Scalars['String']>;
  neighborhood: Scalars['String'];
  city: Scalars['String'];
  state: Scalars['String'];
};

export type NfeCompanyAddressInput = {
  postcode: Scalars['String'];
  street: Scalars['String'];
  number: Scalars['String'];
  complement: Scalars['String'];
  neighborhood: Scalars['String'];
  city: Scalars['String'];
  state: Scalars['String'];
};

export type NfeCompanyCertificate = {
  __typename?: 'NfeCompanyCertificate';
  base64?: Maybe<Scalars['String']>;
  password?: Maybe<Scalars['String']>;
};

export type NfeCompanyCertificateInput = {
  base64?: Maybe<Scalars['String']>;
  password?: Maybe<Scalars['String']>;
};

export type NfeCompanyEmitterCredentials = {
  __typename?: 'NfeCompanyEmitterCredentials';
  consumer_key?: Maybe<Scalars['String']>;
  consumer_secret?: Maybe<Scalars['String']>;
  access_token?: Maybe<Scalars['String']>;
  access_token_secret?: Maybe<Scalars['String']>;
};

export type NfeCompanyEmitterCredentialsInput = {
  consumer_key?: Maybe<Scalars['String']>;
  consumer_secret?: Maybe<Scalars['String']>;
  access_token?: Maybe<Scalars['String']>;
  access_token_secret?: Maybe<Scalars['String']>;
};
// ...
export type NfeCompanyTaxSettings = {
  __typename?: 'NfeCompanyTaxSettings';
  ibpt?: Maybe<Scalars['String']>;
  informacoes_fisco?: Maybe<Scalars['String']>;
  orientacao_danfe?: Maybe<Scalars['String']>;
  nfe_serie?: Maybe<Scalars['Int']>;
  nfe_numero?: Maybe<Scalars['Int']>;
  nfe_numero_dev?: Maybe<Scalars['Int']>;
  nfce_serie?: Maybe<Scalars['Int']>;
  nfce_numero?: Maybe<Scalars['Int']>;
  nfce_numero_dev?: Maybe<Scalars['String']>;
  nfce_id_csc?: Maybe<Scalars['String']>;
  nfce_codigo_csc?: Maybe<Scalars['String']>;
  nfce_id_csc_dev?: Maybe<Scalars['String']>;
  nfce_codigo_csc_dev?: Maybe<Scalars['String']>;
};

export type NfeCompanyTaxSettingsInput = {
  ibpt?: Maybe<Scalars['String']>;
  informacoes_fisco?: Maybe<Scalars['String']>;
  orientacao_danfe?: Maybe<Scalars['String']>;
  nfe_serie?: Maybe<Scalars['Int']>;
  nfe_numero?: Maybe<Scalars['Int']>;
  nfe_numero_dev?: Maybe<Scalars['Int']>;
  nfce_serie?: Maybe<Scalars['Int']>;
  nfce_numero?: Maybe<Scalars['Int']>;
  nfce_numero_dev?: Maybe<Scalars['String']>;
  nfce_id_csc?: Maybe<Scalars['String']>;
  nfce_codigo_csc?: Maybe<Scalars['String']>;
  nfce_id_csc_dev?: Maybe<Scalars['String']>;
  nfce_codigo_csc_dev?: Maybe<Scalars['String']>;
};

export enum NfeCompanyUnit {
  Matriz = 'matriz',
  Filial = 'filial'
}

export enum NfeEnv {
  Production = 'PRODUCTION',
  Homologation = 'HOMOLOGATION'
}
// ...
export type NfePayment = {
  __typename?: 'NfePayment';
  fatura?: Maybe<NfePaymentInvoice>;
  parcelas?: Maybe<Array<Maybe<NfePaymentInstallment>>>;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type UpdateNfeCompanyInput = {
  id: Scalars['ID'];
  ownerName: Scalars['String'];
  corporateName: Scalars['String'];
  email: Scalars['String'];
  phone: Scalars['String'];
  tradeName?: Maybe<Scalars['String']>;
  accounting?: Maybe<Scalars['String']>;
  subdomain?: Maybe<Scalars['String']>;
  logoUrl?: Maybe<Scalars['String']>;
  cpf: Scalars['String'];
  cnpj: Scalars['String'];
  ie: Scalars['String'];
  unit: NfeCompanyUnit;
  address: NfeCompanyAddressInput;
  taxSettings?: Maybe<NfeCompanyTaxSettingsInput>;
  certificate?: Maybe<NfeCompanyCertificateInput>;
  emitterCredentials?: Maybe<NfeCompanyEmitterCredentialsInput>;
};
// ...
export type Zone = Node & {
  __typename?: 'Zone';
  id: Scalars['ID'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
  name: Scalars['String'];
  members: Array<Country>;
  customFields?: Maybe<Scalars['JSON']>;
};
// ...
  export type Fragment = NfeListItemFragment;
// ...
  export type Channels = NonNullable<(NonNullable<NfeCompanyListItemFragment['channels']>)[number]>;
// ...
export namespace NfeCompany {
  export type Fragment = NfeCompanyFragment;
  export type Address = (NonNullable<NfeCompanyFragment['address']>);
  export type TaxSettings = (NonNullable<NfeCompanyFragment['taxSettings']>);
  export type Certificate = (NonNullable<NfeCompanyFragment['certificate']>);
  export type EmitterCredentials = (NonNullable<NfeCompanyFragment['emitterCredentials']>);
  export type Channels = NonNullable<(NonNullable<NfeCompanyFragment['channels']>)[number]>;
}
// ...
export namespace Nfe {
  export type Fragment = NfeFragment;
  export type Info = (NonNullable<NfeFragment['info']>);
  export type Recipient = (NonNullable<NfeFragment['recipient']>);
  export type Payment = (NonNullable<NfeFragment['payment']>);
  export type Fatura = (NonNullable<(NonNullable<NfeFragment['payment']>)['fatura']>);
  export type Parcelas = NonNullable<(NonNullable<(NonNullable<NfeFragment['payment']>)['parcelas']>)[number]>;
  export type Transport = (NonNullable<NfeFragment['transport']>);
  export type Exportation = (NonNullable<NfeFragment['exportation']>);
  export type References = (NonNullable<NfeFragment['references']>);
  export type Ecfs = NonNullable<(NonNullable<(NonNullable<NfeFragment['references']>)['ecfs']>)[number]>;
  export type Company = (NonNullable<NfeFragment['company']>);
  export type Order = (NonNullable<NfeFragment['order']>);
  export type LastOperation = (NonNullable<NfeFragment['lastOperation']>);
  export type Operations = NonNullable<(NonNullable<NfeFragment['operations']>)[number]>;
  export type Products = NonNullable<(NonNullable<NfeFragment['products']>)[number]>;
}

export namespace CreateNfeCompany {
  export type Variables = CreateNfeCompanyMutationVariables;
  export type Mutation = CreateNfeCompanyMutation;
  export type CreateNfeCompany = (NonNullable<CreateNfeCompanyMutation['createNfeCompany']>);
}

export namespace UpdateNfeCompany {
  export type Variables = UpdateNfeCompanyMutationVariables;
  export type Mutation = UpdateNfeCompanyMutation;
  export type UpdateNfeCompany = (NonNullable<UpdateNfeCompanyMutation['updateNfeCompany']>);
}
// ...
export type NfeCompanyFragment = (
  { __typename?: 'NfeCompany' }
  & Pick<NfeCompany, 'id' | 'createdAt' | 'updatedAt' | 'ownerName' | 'corporateName' | 'email' | 'phone' | 'tradeName' | 'accounting' | 'subdomain' | 'logoUrl' | 'cpf' | 'cnpj' | 'ie' | 'unit'>
  & { address: (
    { __typename?: 'NfeCompanyAddress' }
    & Pick<NfeCompanyAddress, 'postcode' | 'street' | 'number' | 'complement' | 'neighborhood' | 'city' | 'state'>
  ), taxSettings?: Maybe<(
    { __typename?: 'NfeCompanyTaxSettings' }
    & Pick<NfeCompanyTaxSettings, 'ibpt' | 'informacoes_fisco' | 'orientacao_danfe' | 'nfe_serie' | 'nfe_numero' | 'nfe_numero_dev' | 'nfce_serie' | 'nfce_numero' | 'nfce_numero_dev' | 'nfce_id_csc' | 'nfce_codigo_csc' | 'nfce_id_csc_dev' | 'nfce_codigo_csc_dev'>
  )>, certificate?: Maybe<(
    { __typename?: 'NfeCompanyCertificate' }
    & Pick<NfeCompanyCertificate, 'base64' | 'password'>
  )>, emitterCredentials?: Maybe<(
    { __typename?: 'NfeCompanyEmitterCredentials' }
    & Pick<NfeCompanyEmitterCredentials, 'consumer_key' | 'consumer_secret' | 'access_token' | 'access_token_secret'>
  )>, channels?: Maybe<Array<Maybe<(
    { __typename?: 'Channel' }
    & ChannelFragmentFragment
  )>>> }
);
// ...
export type CreateNfeCompanyMutationVariables = Exact<{
  input: CreateNfeCompanyInput;
}>;


export type CreateNfeCompanyMutation = (
  { __typename?: 'Mutation' }
  & { createNfeCompany: (
    { __typename?: 'NfeCompany' }
    & NfeCompanyFragment
  ) }
);

export type UpdateNfeCompanyMutationVariables = Exact<{
  input: UpdateNfeCompanyInput;
}>;


export type UpdateNfeCompanyMutation = (
  { __typename?: 'Mutation' }
  & { updateNfeCompany: (
    { __typename?: 'NfeCompany' }
    & NfeCompanyFragment
  ) }
);